==> Notebooks/1-Cleaning_Oscars_csv.ipynb <==
## Cleaning our first oscar's CSV ## 

import pandas as pd

movies = pd.read_csv('/Users/davidusuariocasa/Desktop/oscars.csv') # I open the csv with the Oscar winners from 1928 to 2023

movies


filtered_df = movies[movies['winner'] == True] # I filter by 'True' only to show the winners and discard the nominated who didn't win (those with value 'False')

filtered_df

null_rows = filtered_df[movies.isnull().any(axis=1)] # I do this to see which rows contain null values

finaldf = filtered_df.dropna() # And now I drop all my null values

finaldf # now I have all the winners, without null values

finaldf.reset_index(inplace = True) # Now I reset all indexes in my df

finaldf

columns_to_drop = ['winner', 'index'] # I create a list with the names of the columns I want to remove from my df

finaldf.drop(columns=columns_to_drop, inplace=True) # Now I remove the column 'winner' and 'index'



finaldf # This is my final df

finaldf.to_csv('oscars_cleaned.csv', index=False) # I store my cleaned CSV
