==> Notebooks/2-Web_scrapping.ipynb <==
## Web scraping to enrich our CSV ##

import pandas as pd  
import numpy as np
import re
import pickle
from tqdm import tqdm


from selenium import webdriver     # I import everything I need to do web scrapping with selenium
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options
from selenium.common.exceptions import TimeoutException



# Ignoring the warnings.
import warnings
warnings.filterwarnings('ignore')

# I activate all these options for doing the web scrapping

opciones= Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
#para ocultarme como robot
opciones.add_experimental_option('useAutomationExtension', False)
opciones.add_argument('--start-maximized') 
opciones.add_argument('user.data-dir=selenium')
opciones.add_argument('--incognito

movies = pd.read_csv('/Users/davidusuariocasa/Desktop/Projects/ETL_project/CSVs/Oscars_cleaned.csv') # I open the csv with the CSV I want to iterate from 

## Loop to get each film's 'score' and 'duration' from IMDB ##


driver = webdriver.Firefox()

driver.get("https://www.imdb.com/") # This is the web page we get the info from
driver.implicitly_wait(2)

for index, (value, year) in enumerate(zip(movies['film'], movies['year_film'])): # We iterate through each film (row), and we want to write its title and year
    try:
        
        driver.find_element("css selector", "#suggestion-search").send_keys(f'{value} {str(year)}') # In IMDB's search box, we write each film and its year
        time.sleep(2)

        try:
            
            driver.find_element("css selector", "#react-autowhatever-navSuggestionSearch--item-0 > a:nth-child(1) > div:nth-child(2)").click() # Try to click on the first option that our seach shows
        except:
            movies.loc[index, 'score'] = np.nan # If our movie doesn't have an IMDB profile, we fill our column 'score' and 'duration' with null values, and clear our search box
            movies.loc[index, 'duration'] = np.nan
            driver.find_element("css selector", "#suggestion-search").clear()

            continue  # Continue to the next iteration

        try:
            duration = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[1]/ul/li[3]').text
            movies.loc[index, 'duration'] = duration # If the film has a duration written, take it and assign it to our 'duration' column
        except:
            movies.loc[index, 'duration'] = np.nan # If not, assign a null value to the 'duration' column
            continue  # Continue to the next iteration

        try:
            score = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[2]/div/div[1]/a/span/div/div[2]/div[1]').text
            movies.loc[index, 'score'] = score # If the film has a score written, take it and assign it to our 'score' column
        except:
            movies.loc[index, 'score'] = np.nan # If not, assign a null value to the 'score' column

    except TimeoutException:
        continue  # Go to the next iteration, if it shows an error of Timeout Exception 

driver.quit() # At the end of the for loop, close the Firefox navegator 



## First iteration ##

'''First time we iterated through our 'movie' df, an error appeared at index 260, so we take the first 260 
indexes (because they took the data that we wanted), and store them in a new variable called 'movies_sub1''''

movies_sub1 = movies.iloc[:260] # This is our first subdf, with the data we wanted before the error


## Second iteration ## 

movies_sub2 = movies.iloc[260:] # Our second subdf is composed by the unscrapped rows that didn't work on our first time

movies_sub2 = movies_sub2.reset_index() # We reset the indexes and drop the column 'index'
movies_sub2 = movies_sub2.drop('index', axis=1)
movies_sub2

driver = webdriver.Firefox() 

# Now, we repeat the same iteration as before, but with our subdataframe number 2

driver.get("https://www.imdb.com/") # This is the web page we get the info from
driver.implicitly_wait(2)

for index, (value, year) in enumerate(zip(movies_sub2['film'], movies_sub2['year_film'])): # We iterate through each film (row), and we want to write its title and year
    try:
        
        driver.find_element("css selector", "#suggestion-search").send_keys(f'{value} {str(year)}') # In IMDB's search box, we write each film and its year
        time.sleep(2)

        try:
            
            driver.find_element("css selector", "#react-autowhatever-navSuggestionSearch--item-0 > a:nth-child(1) > div:nth-child(2)").click() # Try to click on the first option that our seach shows
        except:
            movies_sub2.loc[index, 'score'] = np.nan # If our movie doesn't have an IMDB profile, we fill our column 'score' and 'duration' with null values, and clear our search box
            movies_sub2.loc[index, 'duration'] = np.nan
            driver.find_element("css selector", "#suggestion-search").clear()

            continue  # Continue to the next iteration

        try:
            duration = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[1]/ul/li[3]').text
            movies_sub2.loc[index, 'duration'] = duration # If the film has a duration written, take it and assign it to our 'duration' column
        except:
            movies_sub2.loc[index, 'duration'] = np.nan # If not, assign a null value to the 'duration' column
            continue  # Continue to the next iteration

        try:
            score = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[2]/div/div[1]/a/span/div/div[2]/div[1]').text
            movies_sub2.loc[index, 'score'] = score # If the film has a score written, take it and assign it to our 'score' column
        except:
            movies_sub2.loc[index, 'score'] = np.nan # If not, assign a null value to the 'score' column

    except TimeoutException:
        continue  # Go to the next iteration, if it shows an error of Timeout Exception 

driver.quit() # At the end of the for loop, close the Firefox navegator 



movies_sub2 = movies_sub2.iloc[:419]

'''It worked until our index 419 from movies_sub2, so we rewrite the variable
with the rows that worked. This is our second subdataframe'''

## Third iteration ##

movies_sub3 = movies.iloc[679:]
movies_sub3 = movies_sub3.reset_index()
movies_sub3 = movies_sub3.drop('index', axis=1)
movies_sub3

'''Same as before. Our third dataframe is composed by the unscrapped rows that didn't work in our second iteration.
We reset the indexes and drop the 'index' column'''

driver = webdriver.Firefox() 

# Now, we repeat the same iteration as before, but with our subdataframe number 3

driver.get("https://www.imdb.com/") # This is the web page we get the info from
driver.implicitly_wait(2)

for index, (value, year) in enumerate(zip(movies_sub3['film'], movies_sub3['year_film'])): # We iterate through each film (row), and we want to write its title and year
    try:
        
        driver.find_element("css selector", "#suggestion-search").send_keys(f'{value} {str(year)}') # In IMDB's search box, we write each film and its year
        time.sleep(2)

        try:
            
            driver.find_element("css selector", "#react-autowhatever-navSuggestionSearch--item-0 > a:nth-child(1) > div:nth-child(2)").click() # Try to click on the first option that our seach shows
        except:
            movies_sub3.loc[index, 'score'] = np.nan # If our movie doesn't have an IMDB profile, we fill our column 'score' and 'duration' with null values, and clear our search box
            movies_sub3.loc[index, 'duration'] = np.nan
            driver.find_element("css selector", "#suggestion-search").clear()

            continue  # Continue to the next iteration

        try:
            duration = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[1]/ul/li[3]').text
            movies_sub3.loc[index, 'duration'] = duration # If the film has a duration written, take it and assign it to our 'duration' column
        except:
            movies_sub3.loc[index, 'duration'] = np.nan # If not, assign a null value to the 'duration' column
            continue  # Continue to the next iteration

        try:
            score = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[2]/div/div[1]/a/span/div/div[2]/div[1]').text
            movies_sub3.loc[index, 'score'] = score # If the film has a score written, take it and assign it to our 'score' column
        except:
            movies_sub3.loc[index, 'score'] = np.nan # If not, assign a null value to the 'score' column

    except TimeoutException:
        continue  # Go to the next iteration, if it shows an error of Timeout Exception 

driver.quit() # At the end of the for loop, close the Firefox navegator 



movies_sub3 = movies_sub3.iloc[:719] 

'''Our iteration stopped working at index 719, so we rewrite movies_sub3 
with the rows that worked and store it. This is our third subdataframe'''

movies_sub3 # This is our third subdataframe

## Fourth iteration ## 

movies_sub4 = movies.iloc[1398:]
movies_sub4 = movies_sub4.reset_index()
movies_sub4 = movies_sub4.drop('index', axis=1)
movies_sub4

'''We take the unscrapped rows that didn't work on the third iteration, and store it in our fourth subdataframe
We reset indexes and drop the 'index' column'''

driver = webdriver.Firefox()

# Now, we repeat the same iteration as before, but with our subdataframe number 4


driver.get("https://www.imdb.com/") # This is the web page we get the info from
driver.implicitly_wait(2)

for index, (value, year) in enumerate(zip(movies_sub4['film'], movies_sub4['year_film'])): # We iterate through each film (row), and we want to write its title and year
    try:
        
        driver.find_element("css selector", "#suggestion-search").send_keys(f'{value} {str(year)}') # In IMDB's search box, we write each film and its year
        time.sleep(2)

        try:
            
            driver.find_element("css selector", "#react-autowhatever-navSuggestionSearch--item-0 > a:nth-child(1) > div:nth-child(2)").click() # Try to click on the first option that our seach shows
        except:
            movies_sub4.loc[index, 'score'] = np.nan # If our movie doesn't have an IMDB profile, we fill our column 'score' and 'duration' with null values, and clear our search box
            movies_sub4.loc[index, 'duration'] = np.nan
            driver.find_element("css selector", "#suggestion-search").clear()

            continue  # Continue to the next iteration

        try:
            duration = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[1]/ul/li[3]').text
            movies_sub4.loc[index, 'duration'] = duration # If the film has a duration written, take it and assign it to our 'duration' column
        except:
            movies_sub4.loc[index, 'duration'] = np.nan # If not, assign a null value to the 'duration' column
            continue  # Continue to the next iteration

        try:
            score = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[2]/div/div[1]/a/span/div/div[2]/div[1]').text
            movies_sub4.loc[index, 'score'] = score # If the film has a score written, take it and assign it to our 'score' column
        except:
            movies_sub4.loc[index, 'score'] = np.nan # If not, assign a null value to the 'score' column

    except TimeoutException:
        continue  # Go to the next iteration, if it shows an error of Timeout Exception 

driver.quit() # At the end of the for loop, close the Firefox navegator 



movies_sub4 = movies_sub4[:748] 
movies_sub4

'''Our iteration stopped working at index 748, so we rewrite movies_sub4 
with the rows that worked and store it. This is our fourth subdataframe'''

## Fifth and last iteration ##

movies_sub5 = movies.iloc[2146:]
movies_sub5 = movies_sub5.reset_index()
movies_sub5 = movies_sub5.drop('index', axis=1)
movies_sub5 # These are the remaining rows to iterate

driver = webdriver.Firefox()

# Now, we repeat the same iteration as before, but with our subdataframe number 5


driver.get("https://www.imdb.com/") # This is the web page we get the info from
driver.implicitly_wait(2)

for index, (value, year) in enumerate(zip(movies_sub5['film'], movies_sub5['year_film'])): # We iterate through each film (row), and we want to write its title and year
    try:
        
        driver.find_element("css selector", "#suggestion-search").send_keys(f'{value} {str(year)}') # In IMDB's search box, we write each film and its year
        time.sleep(2)

        try:
            
            driver.find_element("css selector", "#react-autowhatever-navSuggestionSearch--item-0 > a:nth-child(1) > div:nth-child(2)").click() # Try to click on the first option that our seach shows
        except:
            movies_sub5.loc[index, 'score'] = np.nan # If our movie doesn't have an IMDB profile, we fill our column 'score' and 'duration' with null values, and clear our search box
            movies_sub5.loc[index, 'duration'] = np.nan
            driver.find_element("css selector", "#suggestion-search").clear()

            continue  # Continue to the next iteration

        try:
            duration = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[1]/ul/li[3]').text
            movies_sub5.loc[index, 'duration'] = duration # If the film has a duration written, take it and assign it to our 'duration' column
        except:
            movies_sub5.loc[index, 'duration'] = np.nan # If not, assign a null value to the 'duration' column
            continue  # Continue to the next iteration

        try:
            score = driver.find_element('xpath', '/html/body/div[2]/main/div/section[1]/section/div[3]/section/section/div[2]/div[2]/div/div[1]/a/span/div/div[2]/div[1]').text
            movies_sub5.loc[index, 'score'] = score # If the film has a score written, take it and assign it to our 'score' column
        except:
            movies_sub5.loc[index, 'score'] = np.nan # If not, assign a null value to the 'score' column

    except TimeoutException:
        continue  # Go to the next iteration, if it shows an error of Timeout Exception 

driver.quit() # At the end of the for loop, close the Firefox navegator 



movies_sub5 # This is our fifth and last subdataframe

 ## Concatenation of all subdataframes ## 

df_final = pd.concat([movies_sub1, movies_sub2, movies_sub3, movies_sub4, movies_sub5], ignore_index=True)

'''We concatenate all subdataframes and store them in a variable called df_final'''


df_final

df_final.to_csv('scrapped_csv', index=False) # We save our final df as a CSV
