==> Notebooks/3-Cleaning_scrapped_csv.ipynb <==
## Cleaning our scrapped Oscar's CSV ## 

import os

ruta_csv = os.path.abspath('nombre_del_archivo.csv')
print(ruta_csv)

import pandas as pd


scrapped = pd.read_csv('/Users/davidusuariocasa/Desktop/Projects/ETL_project/Data/scrapped_not_clean.csv'
)


scrapped = scrapped.dropna() # First, we remove all columns that contain at least one null value

scrapped = scrapped.reset_index() # Now we reset the indexes

scrapped = scrapped.drop('index', axis=1) # And finally we remove the column 'index'


## Converting our 'duration' column to minutes ##

import re # We import the regex library

# We want to convert the values in our 'duration' column to minutes, so we create a function to do so

def convert_duration_to_minutes(value): 
    hours = re.findall(r'\d+h', value) # we use regex to get all values followed by an 'h' (hours) and store them
    minutes = re.findall(r'\d+min', value) # the same, but with mins

    hours = int(hours[0].rstrip('h')) if hours else 0 # If regex finds hours, we take out the 'h' and store the number
    minutes = int(minutes[0].rstrip('min')) if minutes else 0 # If regex finds mins, we take 'mins' and store the number

    total_minutes = hours * 60 + minutes # since we want everything in mins, we multiply the hours by 60 and add the remaining minutes

    return total_minutes # our function returns the total mins





scrapped['duration'] = scrapped['duration'].apply(convert_duration_to_minutes) # we apply the function to our 'duration' column


scrapped

## Changing our 'score' column ## 

def clean_rating(value):
    parts = value.split('\n/')
    if len(parts) == 2:
        try:
            cleaned_value = parts[0].replace(',', '.').strip()  # Reemplazar comas por puntos y eliminar espacios en blanco
            return float(cleaned_value)
        except ValueError:
            return None
    else:
        return None



scrapped['score'] = scrapped['score'].apply(clean_rating) # we apply the function to our 'score' column


scrapped

## Changing our 'category' column ## 

# All the values of our 'category' column are in capital letters. We want them to be in lowercase, so we create a function

def capitalize_words(value):
    words = value.split()  # Divide the value in words
    capitalized_words = []  # List to store the capitalized words

    for word in words:
        # Capitalize the first letter of each word
        capitalized_word = word[0].upper() + word[1:].lower()
        capitalized_words.append(capitalized_word)  # Add the capitalized word to our list 
    # Now we join the capitalized words in a single string
    result = ' '.join(capitalized_words)
    return result



scrapped['category'] = scrapped['category'].apply(capitalize_words) # we apply the function to our 'category' column


## Changing our columns names ## 


# Finally we replace the values of my columns with the values from the 'new_names' list

new_names = ['Film year', 'Ceremony year', 'Ceremony number', 'Winning category', 'Winner name', 'Film title', 'Film Duration', 'Film score']

scrapped.columns = new_names


scrapped # This is our final dataframe

scrapped.to_csv('scrapped_clean.csv', index=False) # We save our dataframe in CSV format
