==> Notebooks/4-SQL.ipynb <==
## SQL ALCHEMY ## 

!pip install sqlalchemy==1.3.24
!pip install pymysql
!pip install mysql-connector-python


import pandas as pd
import numpy as np
import sqlalchemy as alch

from getpass import getpass
from tqdm import tqdm 

password = getpass("Contraseña de MySQL: ")


clean = pd.read_csv('/Users/davidusuariocasa/Desktop/Projects/ETL_project/Data/scrapped_clean.csv')

# We store our database name in a variable
database = "Movies"

# We connect with SQL
conexion = f"mysql+pymysql://root:{password}@localhost/{database}"

engine = create_engine(conexion)

clean # This is the df that we want to import to SQL

for index, row in clean.iterrows():
    try:
        query = f"""
        INSERT INTO clientes (opp_id, name, email, city) 
        VALUES ('{row["Opportunity ID"]}', '{row["Name"]}', '{row["Email"]}', '{row["City"]}');
        """
        print(query)
        engine.execute(query)
    except:
        print(row["Opportunity ID"])

## 'Films' table ## 

for index, row in clean.iterrows():
    try:
        engine.execute(f"""
            INSERT INTO Films ('Film year', 'Film title', 'Film duration') VALUES
            ('{row["Film year"]}', '{row["Film title"]}', '{row["Film Duration"]}');""")
        engine.commit()  # Confirmar los cambios y hacer un commit
    except:
        pass


## 'Categories' table ##

for index, row in clean.iterrows():

    try: 

        engine.execute(f"""
            INSERT INTO Categories (Winning category) VALUES
            ('{row["Winning category"]}');""")
    
    except: 

        pass

## 'Winners' table ## 

for index, row in clean.iterrows():
    
    try:
        
        engine.execute(f"""
            INSERT INTO Winners ('Ceremony year', 'Ceremony number', 'Winner name', 'Film score') VALUES
            ('{row["Ceremony year"]}', '{row["Ceremony number"]}', '{row["Winner name"]}', '{row["Film score"]}');""")
        
        engine.commit()  # Confirmar los cambios y hacer un commit
        
    except:
        
        pass
